+++ crates/module-system/module-implementations/integration-tests/src/visible_hash/soft_confirmations.rs
@@ -16,7 +16,7 @@ generate_bare_runtime! {
     modules: [visible_hash_module: TestVisibleHashModule<S>],
     operating_mode: sov_modules_api::OperatingMode::Optimistic,
     minimal_genesis_config_type: sov_test_utils::runtime::genesis::optimistic::MinimalOptimisticGenesisConfig<S>,
-    impl_hooks: [ApplyBatchHooks, TxHooks],
+    impl_hooks: [ApplyBatchHooks, KernelSlotHooks, TxHooks],
     runtime_trait_impl_bounds: [],
     kernel_type: sov_kernels::soft_confirmations::SoftConfirmationsKernel<'a, S>
 }